NLP_ContentIntelligenceAgency_Phase2/src/pyproject.toml
[tool.poetry]
name        = "project-nlp8"
version     = "0.1.0"
description = ""
authors     = ["AlexandruMic236535 <144055838+AlexandruMic236535@users.noreply.github.com>"]
readme      = "README.md"
packages    = [{ include = "project_nlp8", from = "src" }]

[tool.poetry.dependencies]
python          = ">=3.9,<3.11"

# core libs
numpy           = "1.23.5"
pandas          = ">=2.2.3,<3.0.0"
matplotlib      = ">=3.5.3,<4.0.0"
seaborn         = ">=0.13.2,<0.14.0"
scikit-learn    = ">=1.6.1,<2.0.0"

# ML / DL runtime
transformers                 = ">=4.51.3,<5.0.0"

# Web API stack
fastapi         = "^0.111.0"
starlette       = "^0.37.0"
python-multipart = "^0.0.7"

# Docs / misc
sphinx-rtd-theme = ">=1.3.0,<2.0.0"
pre-commit       = ">=4.2.0,<5.0.0"

# AzureML
azureml-core = "^1.55.0"
azure-ai-ml = "^1.16.0"
azure-identity = "^1.23.0"
marshmallow = "3.20.2"

[tool.poetry.group.dev.dependencies]
pytest      = "^8.4"
pytest-cov  = "^6.1.1"
flake8      = "^7.2.0"
black       = "^25.1"
isort       = "^6.0.1"
autoflake   = "^2.3.1"
contractions = "^0.1.72"
emoji        = "^2.1.0"
textblob     = "^0.17.1"
typer        = "^0.12"
uvicorn      = { extras = ["standard"], version = "^0.29.0" }
pytest-xdist = "^3.7.0"

[build-system]
requires      = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

